Fix - systemd::mount mishandles $enable
This parameter is supposed to control whether the mount is to be brought
up during boot but this was not happening because the unit was never
installed to any target.  The multi-user.target was chosen here as that
seems where most Fedora services land.  (I tried basic.target but the
test host would boot without the mounts occurring.  Investigation showed
that systemd was blocking the mounts because they would have created
dependency cycles, likely due to some $mnt_requires that was passed
along.)