Fix - systemd::unit not restarted after content change
If a unit file's content is changed the unit should be restarted
immediately.  This does seem scary for a service where high-availability
is expected, but is arguably better than having that services behavior
change at some indeterminate future time where it becomes increasingly
difficult to trace the behavioral change back to the cause that effected
it.