Drop - Define[systemd::mount]
Recent systemd.mount(5) pages state:

    In general, configuring mount points through /etc/fstab is the
    preferred approach.

This definition has had difficulty in matching the behavior of
systemd-escape in the past.  With the Puppet Master/Server running on
a new enough OS it would have been possible to match it exactly through
the systemd-escape command that is now provided.  However, given the
above recommendation, it seems best to simply drop this altogether and
resort to using Puppet's traditional mount type that affects /etc/fstab.