Bug - systemd_escaped_mount_path() may include newline
In some versions of Puppet this was not an issue in actual use because
the value would be stripped of newline (or possibly any trailing white
space) characters.  Whether that happened via the File resource or
elsewhere is entirely unknown.  For certain, this issue was evident with
puppet-4.10.1-2.fc27.noarch.

Signed-off-by: John Florian <jflorian@doubledog.org>